Added a block detection executable
block-in-image will indicate if a block is present within
an image passed in via command-line arguments. Currently,
the source will fail to compile with a linker error since
block_in_image within block-detector is not implemented.
This will continue to fail to build until a stub for
block_in_image is created.